chore: rip dead code, dedup copy-paste utilities (code review 2026-06-12 Wave 3a)
Deletions (all verified zero production reachability):
- Scorecard surface: /api/scorecard/submit (orphan POST endpoint — the
  /scorecard page has 301'd home since 2026-05-04), src/scripts/
  scorecard.ts, src/lib/scorecard/, scorecardReportEmailHtml,
  scorecard-template.tsx + renderScorecardReport, scorecard-scoring test
- Parked partner-nurture modules (prompt + schema; no runtime consumer,
  carried retired revenue-band anchoring; #714 rebuilds fresh)
- Dead portal components: PortalHomeDashboard.astro + ActionCard.astro
  (only importers were dev-gated galleries) + design-preview/portal-home
- Dead exports: quoteSentEmailHtml (test-only zombie) + its source-mirror
  test block, deprecated SESSION_DURATION_MS + auth/index re-export,
  two orphaned JSDoc blocks in email/templates.ts

Dedup:
- trimString/isValidEmail/escapeHtml/jsonResponse: 3 byte-identical
  private copies (intake, intake/send, reserve-helpers) -> canonical
  src/lib/api/helpers.ts; reserve-helpers re-exports for compat
- ALLOWED_INTERESTS: duplicated 13-value set in book.astro and
  intake/send.ts -> shared export in lib/booking/config.ts

Also:
- assessment-to-quote SYSTEM_PROMPT: drop '750K to 5M annual revenue'
  anchoring (retired band, ADR 0003 / no-revenue-band standard) from the
  live quote-generation prompt
- CLAUDE.md: crane_sod -> crane_sos (stale tool name)

Reviewed-and-declined: the parseLineItems 'duplicate' (assessment-to-
quote vs db/quotes) — different contracts (LLM-output parsing with fence
stripping vs persisted-JSON parsing); not consolidatable without
behavior change.
