--- src/components/portal/ActionCard.astro
@@ -1,53 +1,0 @@
----
-import MoneyDisplay from './MoneyDisplay.astro'
-
-interface Props {
-  pillLabel: string
-  amountCents?: number | null
-  amountLabel?: string | null
-  ctaLabel: string
-  ctaHref: string
-  ctaSubtext?: string | null
-}
-
-const { pillLabel, amountCents, amountLabel, ctaLabel, ctaHref, ctaSubtext } = Astro.props
----
-
-<section
-  class="border-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-background)]"
-  aria-label={pillLabel}
->
-  <div
-    class="flex items-center gap-3 bg-[color:var(--ss-color-text-primary)] text-[color:var(--ss-color-primary)] px-5 py-2.5 font-['Archivo_Narrow'] text-label uppercase tracking-[0.18em] font-bold"
-  >
-    {pillLabel}
-  </div>
-  <div class="px-5 py-5">
-    {
-      typeof amountCents === 'number' && (
-        <>
-          <MoneyDisplay amountCents={amountCents} size="hero" />
-          {amountLabel && (
-            <p class="mt-3 font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] font-semibold text-[color:var(--ss-color-text-secondary)]">
-              {amountLabel}
-            </p>
-          )}
-        </>
-      )
-    }
-  </div>
-  <a
-    href={ctaHref}
-    class="flex items-center justify-between border-t-[3px] border-[color:var(--ss-color-text-primary)] bg-[color:var(--ss-color-primary)] hover:bg-[color:var(--ss-color-primary-hover)] px-5 py-4 font-['Archivo'] text-sm font-bold uppercase tracking-[0.08em] text-[color:var(--ss-color-background)] transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[color:var(--ss-color-action)] focus-visible:ring-inset"
-  >
-    <span>{ctaLabel}</span>
-    <span aria-hidden="true" class="font-black">→</span>
-  </a>
-  {
-    ctaSubtext && (
-      <p class="px-5 py-3 border-t-[2px] border-[color:var(--ss-color-text-primary)] font-['Archivo_Narrow'] text-label uppercase tracking-[0.14em] text-[color:var(--ss-color-text-secondary)]">
-        {ctaSubtext}
-      </p>
-    )
-  }
-</section>
