+++ src/lib/auth/index.ts
@@ -20,7 +20,6 @@ export {
   buildSessionCookie,
   parseSessionToken,
   SESSION_COOKIE_NAME,
-  SESSION_DURATION_MS,
 } from './session'
 export type { SessionData, UserRole } from './session'
 export { asUserRole } from './session'
