--- src/pages/design-preview/portal-home.astro
@@ -1,34 +1,0 @@
----
-// src/pages/design-preview/portal-home.astro
-// Dev-only preview. Production exclusion via import.meta.env.DEV guard.
-
-import '../../styles/global.css'
-import fixtureData from './portal-home.fixture.json'
-import PortalHomeDashboard from '../../components/portal/PortalHomeDashboard.astro'
-
-if (!import.meta.env.DEV) {
-  return Astro.redirect('/404')
-}
----
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="robots" content="noindex, nofollow" />
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Archivo:wght@400;500;600;700;900&family=Archivo+Narrow:wght@400;500;600;700&family=JetBrains+Mono:wght@400;500&display=swap"
-      rel="stylesheet"
-    />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:wght,FILL@100..700,0..1&display=swap"
-      rel="stylesheet"
-    />
-    <title>[preview] PortalHomeDashboard</title>
-  </head>
-  <body class="min-h-screen bg-[color:var(--ss-color-background)]">
-    <PortalHomeDashboard {...fixtureData} />
-  </body>
-</html>
